Required changes in getToolVersion
