+++ tests/internal/e2e/test/specs/vcast.create_second_test_1.test.ts
@@ -16,7 +16,7 @@ import {
   updateTestID,
 } from "../test_utils/vcast_utils";
 import { TIMEOUT } from "../test_utils/vcast_utils";
-import { getToolVersion } from "../../../../unit/utils";
+import { getToolVersion } from "../../../../unit/getToolversion";
 
 describe("vTypeCheck VS Code Extension", () => {
   let bottomBar: BottomBarPanel;
